New cache clear strategy (#433)
* New cache clear strategy that attempts to first clear out short term cache files before clearing long term (think subset variable precomputation) files

* Don't clear dataset cache during processDataset if the taskThread is processingSubset

* Clearing cache at the end of createSubsetVariablesTable caused a different race condition in some of the tests (when the test is making requests). While the timing of this would be unlikely to be seen in production, I'm removing the cache clear.

* Lengthen the Jetty startup time, I believe the errors are coming from the server not getting everything loaded.

+++ WEB-INF/classes/gov/noaa/pfel/erddap/Erddap.java
@@ -24973,9 +24973,11 @@ widgets.select("frequencyOption", "", 1, frequencyOptions, frequencyOption, "") 
     // add new info to categoryInfo
     addRemoveDatasetInfo(true, this.categoryInfo, dataset);
 
-    // clear the dataset's cache
-    // since axis values may have changed and "last" may have changed
-    File2.deleteAllFiles(dataset.cacheDirectory());
+    if (!dataset.isProcessingSubset()) {
+      // clear the dataset's cache
+      // since axis values may have changed and "last" may have changed
+      File2.deleteAllFiles(dataset.cacheDirectory());
+    }
 
     String change = dataset.changed(oldDataset);
     if (change.isEmpty() && dataset instanceof EDDTable) {
